Extract ProtectedRoute component
