+++ src/components/auth/ProtectedRoute.tsx
@@ -1,0 +1,24 @@
+import { useEffect } from 'react';
+import { Navigate, useLocation } from 'react-router-dom';
+import { useAuth } from '../../contexts/AuthContext';
+
+export function ProtectedRoute(props: { children: React.ReactNode }): JSX.Element {
+  const { user, loading, isDemo } = useAuth();
+  const location = useLocation();
+
+  useEffect(() => {
+    if (!user && !loading && !isDemo) {
+      sessionStorage.setItem('redirectUrl', location.pathname);
+    }
+  }, [user, loading, isDemo, location]);
+
+  if (loading) {
+    return <div className="flex h-screen items-center justify-center">Loading...</div>;
+  }
+
+  if (!user && !isDemo) {
+    return <Navigate to="/login" replace />;
+  }
+
+  return <>{props.children}</>;
+}
